rename and improve crap.c
